requirements.txt
# Flask==2.3.3
# SQLAlchemy==2.0.20
# PyMySQL==1.1.0
# PyJWT==2.8.0
# python-dotenv==1.0.0
# twilio==8.5.0
# requests==2.31.0
# Werkzeug==2.3.7

Flask==3.0.2
SQLAlchemy==2.0.27
PyMySQL==1.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
Flask-Cors==4.0.0
sendgrid==6.11.0
twilio==8.12.0
requests==2.31.0
Werkzeug==3.0.1
